Ingredients Class
So I added the class, the addingredient method that goes to a textfile,
and i checked if it works. It does. Now i'm trying to write the JUnit
tests that will deal with the class ! as soon as we have those tests, we
will be able to copy paste them to other classes, with minor changes